+++ cli/csstools-cli/src/plugins/postcss-unset-value.ts
@@ -1,0 +1,18 @@
+import plugin from '@csstools/postcss-unset-value';
+import { cli, helpTextLogger } from '@csstools/base-cli';
+
+export default function postcssUnsetValue() {
+	cli(
+		plugin,
+		['preserve'],
+		helpTextLogger(
+			'@csstools/cli postcss-unset-value',
+			'PostCSS Stepped Unset Value',
+			'Use the unset keyword in CSS.',
+			{
+				preserve: false,
+			},
+		),
+		false,
+	);
+}